made the schema work but still need work on the seeder
Updated the seeder to have it take the constraints and have a schema
that works.

Only issue right now is the concurrency of the tables being created. If
we create them in the order that they are stored, one after the other,
they create the complete db.

PS: added to gitignore some logs