Finalized schema and have now the seed up and seed down working
Finalized schema and have now the seed up and seed down working.

Need some other people to test it and then it is good to go. This means
just pulling the code, having the mysql db ready, configuring the
config file and running npm start DB=up and npm start DB=down and
veryfing that the tables are created and dropped.